Refs #3326. Refs #3504. Fixes #4369. Syncs multirec to trunk.
This touches on many aspects of recording, for DVB Recorders in particular.
 * Adds PID priorities to the MPEGStreamHandler. This allows a DTV recorder to
   place a higher priority on recording audio and visual streams over other streams.
   This is particularly useful for DTV hardware which have a fixed number of hardware
   PID filters. (A similar mechanism already existed in DVBRecorder, but it didn't
   work in multirec, and this more general approach allows other recorders to use this.)
 * Adds input groups. This is used by the DVB Recorder to implement multiple recordings
   per transport. Multiple recorders are created referencing the same device, but they
   are placed in the same input groups so that they are forced to all record on the same
   transport. This can also allow two or more devices to share a single STB, and could
   be extended to allow multiple capture cards to share a single satellite dish on a
   rotor.
   * Removes analog child cards for DVB Recorders (can be simulated with input groups)
     Note: you will need to manually add both cards to the same named input group.
 * DVB/QAM Audio stream support
   * Adds the ability to play audio only mpeg files to the NuppelVideoRecorder.
     Note: this allows all DTV recorders to support "Radio" streams.
   * Removes the dummy video from the DVBRecorder, not needed anymore
 * Adds a new input switching mode which cycles through video sources, instead
   of actual cards. The card switching mode keymap is now undefined by default,
   and the video source switching uses it's "Y" keybinding. Unbind "NEXTSOURCE"
   and bind "NEXTCARD" instead for the old behaviour.
 * Adds the ability to record multiple streams in a TS transport at once with DVB
   cards. Just delete your DVB cards in mythtv-setup and then as you recreate each
   one go into "Recording Options" and set "Max Recordings" to 2 or more. We recommend
   a reasonable number here, as you still have limited hard drive, CPU, etc.

There is still a known problem when using input groups with a non-DVB recorder:
 * The firewire recorder can fail to make recordings if its input is in an input
   group with another recorder's input. Please run the backend with '-v channel,record'
   if you have this configuration and attach the log to #3326 if this happens.
   I found the probable location of the problem added some debugging last week.

We're also very interested in any new problems that are encountered when this in
run with different hardware. About 10 people have tested this, so it is relatively
well tested, but it is a very large change to the way inputs are handled, and
to the DVB Recorder in particular.

This was worked on by Janne, Shane, and myself with help from David and Bruce with
scheduler suggestions and testing, and the help of many others with additonal testing.



git-svn-id: http://svn.mythtv.org/svn/trunk@15437 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/analogsignalmonitor.cpp
@@ -14,7 +14,7 @@
 #define LOC_ERR QString("AnalogSM, Error: ").arg(channel->GetDevice())
 
 AnalogSignalMonitor::AnalogSignalMonitor(int db_cardnum, Channel *_channel,
-                                         uint _flags, const char *_name) :
+                                         uint64_t _flags, const char *_name) :
     SignalMonitor(db_cardnum, _channel, _flags, _name),
     usingv4l2(false)
 {
